stress-dev: cast -1 to unsigned int

+++ stress-dev.c
@@ -2616,7 +2616,7 @@ static void stress_dev_console_linux(
 			 * of scancode and keycode for scancode < 89
 			 */
 			(void)memset(&bad_arg, 0, sizeof(bad_arg));
-			bad_arg.scancode = -1;
+			bad_arg.scancode = (unsigned int)-1;
 			bad_arg.keycode = 2;
 
 			ret = ioctl(fd, KDSETKEYCODE, &bad_arg);
